city_housing_index/requirement.txt
django==2.1.*
requests==2.23.0
beautifulsoup4==4.9.1
fake-useragent==0.1.11
djangorestframework==3.11.0
markdown==3.2.2
django-filter==2.3.0
openpyxl==3.0.4
gunicorn==20.0.4
mysqlclient==2.0.1
django-celery==3.3.1
redis==2.10.6

pycryptodome==3.9.8
pillow==7.2.0
sentry-sdk==0.16.1
jsonfield==3.1.0

xlrd==1.2.0
openpyxl==3.0.4
sklearn==0.0